Deduplicate loop analysis
Also remove some risky truthy falsy comparisons
