+++ psalm-baseline.xml
@@ -469,18 +469,10 @@
       <code><![CDATA[$source_pos = strpos($source, '*')]]></code>
     </RiskyTruthyFalsyComparison>
   </file>
-  <file src="src/Psalm/Internal/Analyzer/Statements/Block/DoAnalyzer.php">
-    <RiskyTruthyFalsyComparison>
-      <code><![CDATA[$do_context->branch_point]]></code>
-    </RiskyTruthyFalsyComparison>
-  </file>
   <file src="src/Psalm/Internal/Analyzer/Statements/Block/ForAnalyzer.php">
     <ArgumentTypeCoercion>
       <code><![CDATA[$stmt->cond]]></code>
     </ArgumentTypeCoercion>
-    <RiskyTruthyFalsyComparison>
-      <code><![CDATA[$for_context->branch_point]]></code>
-    </RiskyTruthyFalsyComparison>
   </file>
   <file src="src/Psalm/Internal/Analyzer/Statements/Block/ForeachAnalyzer.php">
     <ConflictingReferenceConstraint>
@@ -496,7 +488,6 @@
       <code><![CDATA[!$var_comment->var_id]]></code>
       <code><![CDATA[!$var_comment->var_id]]></code>
       <code><![CDATA[$calling_type_params]]></code>
-      <code><![CDATA[$foreach_context->branch_point]]></code>
       <code><![CDATA[$generic_storage->template_types]]></code>
       <code><![CDATA[$statements_analyzer->getTemplateTypeMap()]]></code>
       <code><![CDATA[$var_comment->line_number]]></code>
@@ -526,7 +517,6 @@
   <file src="src/Psalm/Internal/Analyzer/Statements/Block/IfElseAnalyzer.php">
     <RiskyTruthyFalsyComparison>
       <code><![CDATA[$context->branch_point]]></code>
-      <code><![CDATA[$else_context->branch_point]]></code>
       <code><![CDATA[$else_context->branch_point]]></code>
       <code><![CDATA[$if_scope->assigned_var_ids]]></code>
       <code><![CDATA[$if_scope->new_vars]]></code>
@@ -558,22 +548,11 @@
       <code><![CDATA[$traverser->traverse([$switch_condition])[0]]]></code>
     </PossiblyUndefinedIntArrayOffset>
     <RiskyTruthyFalsyComparison>
-      <code><![CDATA[$case_context->branch_point]]></code>
       <code><![CDATA[$nested_or_options]]></code>
       <code><![CDATA[$switch_var_id]]></code>
       <code><![CDATA[$switch_var_id]]></code>
       <code><![CDATA[$switch_var_id]]></code>
       <code><![CDATA[$type_statements]]></code>
-    </RiskyTruthyFalsyComparison>
-  </file>
-  <file src="src/Psalm/Internal/Analyzer/Statements/Block/TryAnalyzer.php">
-    <RiskyTruthyFalsyComparison>
-      <code><![CDATA[$try_context->branch_point]]></code>
-    </RiskyTruthyFalsyComparison>
-  </file>
-  <file src="src/Psalm/Internal/Analyzer/Statements/Block/WhileAnalyzer.php">
-    <RiskyTruthyFalsyComparison>
-      <code><![CDATA[$while_context->branch_point]]></code>
     </RiskyTruthyFalsyComparison>
   </file>
   <file src="src/Psalm/Internal/Analyzer/Statements/DeclareAnalyzer.php">
