+++ src/Psalm/Internal/Analyzer/Statements/Block/DoAnalyzer.php
@@ -40,8 +40,8 @@ final class DoAnalyzer
 
         $codebase = $statements_analyzer->getCodebase();
 
-        if ($codebase->alter_code) {
-            $do_context->branch_point = $do_context->branch_point ?: (int) $stmt->getAttribute('startFilePos');
+        if ($codebase->alter_code && $do_context->branch_point === null) {
+            $do_context->branch_point = (int) $stmt->getAttribute('startFilePos');
         }
 
         $loop_scope = new LoopScope($do_context, $context);
