+++ src/Psalm/Internal/Scope/LoopScope.php
@@ -40,7 +40,7 @@ final class LoopScope
     public array $vars_possibly_in_scope = [];
 
     /**
-     * @var array<string, bool>
+     * @var array<string, bool|int>
      */
     public array $protected_var_ids = [];
